- added shared tests for MediaElement and WebAudioMedia backend. - updated changelog and added examples.
