+++ src/webaudio-media.js
@@ -1,12 +1,14 @@
 import MediaElement from './mediaelement';
 
 /**
- * WebAudioMedia backend: allows to load audio as HTML5 audio tag and use it with WebAudio backend.
+ * WebAudioMedia backend: allows to load audio as HTML5 audio tag and use it with API Web Audio.
  * Setting the WebAudioMedia backend, there is the possibility to load audio of big dimensions, using the API Web Audio features.
- * In this way, filters and other functionalities can be performed as with WebAudio backend without decoding
- * internally audio data, that caused crashing of the browser.
+ * The audio to load is an HTML5 audio tag, so you have to use the same methods of MediaElement backend for loading and playback.
+ * In this way, the audio resource is not loaded entirely from server, but in ranges, since you load an HTML5 audio tag.
+ * In this way, filters and other functionalities can be performed like with WebAudio backend, but without decoding
+ * internally audio data, that caused crashing of the browser. You have to give also peaks, so the audio data are not decoded.
  *
- * @since Version 3.2.0
+ * @since 3.2.0
  */
 export default class WebAudioMedia extends MediaElement {
     /**
@@ -48,7 +50,7 @@ export default class WebAudioMedia extends MediaElement {
     /**
      * Create MediaElementSource node
      *
-     * @since Version 3.2.0
+     * @since 3.2.0
      * @param {HTMLMediaElement} mediaElement HTML5 Audio to load
      */
     createMediaElementSource(mediaElement) {
